Refactor: Created Graph subfolder, Inserted 2 graph hooks, Fixed useConnectedWeb3 context imports when importing from index file everywhere
