+++ app/src/App.tsx
@@ -9,8 +9,8 @@ import 'sanitize.css'
 import { HeaderNoRouter } from './components/common/layout/header'
 import { Main } from './components/main'
 import SettingsViewContainer from './components/settings/settings_view'
+import { ConnectedWeb3 } from './contexts'
 import { ApolloProviderWrapper } from './contexts/Apollo'
-import { ConnectedWeb3 } from './hooks'
 import balanceReducer from './store/reducer'
 import { ThemeProvider } from './theme'
 import { GlobalStyle } from './theme/global_style'
