+++ app/src/components/common/layout/header/index.tsx
@@ -6,7 +6,7 @@ import styled, { css } from 'styled-components'
 
 import { Network } from '../..'
 import { Logo, STANDARD_DECIMALS } from '../../../../common/constants'
-import { useConnectedWeb3Context } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
 import { networkIds } from '../../../../util/networks'
 import { formatBigNumber } from '../../../../util/tools'
 import { ExchangeType } from '../../../../util/types'
