+++ app/src/components/market/common/additional_market_data/index.tsx
@@ -2,7 +2,8 @@ import React, { DOMAttributes, useEffect, useState } from 'react'
 import ReactTooltip from 'react-tooltip'
 import styled from 'styled-components'
 
-import { useConnectedWeb3Context, useRealityLink } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useRealityLink } from '../../../../hooks'
 import { CompoundService } from '../../../../services/compound_service'
 import { networkIds } from '../../../../util/networks'
 import { Arbitrator, CompoundTokenType, KlerosItemStatus, KlerosSubmission, Token } from '../../../../util/types'
