+++ app/src/components/market/common/currency_selector/index.tsx
@@ -2,7 +2,8 @@ import { BigNumber } from 'ethers/utils'
 import React, { useEffect } from 'react'
 import styled, { css } from 'styled-components'
 
-import { ConnectedWeb3Context, useTokens } from '../../../../hooks'
+import { ConnectedWeb3Context } from '../../../../contexts'
+import { useTokens } from '../../../../hooks'
 import { formatBigNumber } from '../../../../util/tools'
 import { Token } from '../../../../util/types'
 import { Dropdown, DropdownItemProps, DropdownPosition } from '../../../common/form/dropdown'
