+++ app/src/components/market/common/history_table/index.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { useConnectedWeb3Context } from '../../../../hooks'
-import { FpmmTradeDataType } from '../../../../hooks/useGraphFpmmTransactionsFromQuestion'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { FpmmTradeDataType } from '../../../../hooks/Graph/useGraphFpmmTransactionsFromQuestion'
 import { getTxHashBlockExplorerURL } from '../../../../util/networks'
 import { formatBigNumber, formatHistoryDate, formatHistoryUser } from '../../../../util/tools'
 import { ButtonRound } from '../../../button'
