+++ app/src/components/market/common/market_data/index.tsx
@@ -4,7 +4,8 @@ import momentTZ from 'moment-timezone'
 import React, { DOMAttributes, useEffect, useState } from 'react'
 import styled from 'styled-components'
 
-import { useConnectedWeb3Context, useSymbol, useTokens } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useSymbol, useTokens } from '../../../../hooks'
 import { CompoundService } from '../../../../services'
 import { getNativeAsset, getNativeCompoundAsset, getToken } from '../../../../util/networks'
 import { formatBigNumber, formatDate, formatToShortNumber, getBaseTokenForCToken } from '../../../../util/tools'
