+++ app/src/components/market/common/outcome_table/index.tsx
@@ -3,7 +3,8 @@ import { BigNumber } from 'ethers/utils'
 import React, { useCallback } from 'react'
 import styled from 'styled-components'
 
-import { useConnectedWeb3Context, useSymbol } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useSymbol } from '../../../../hooks'
 import { getOutcomeColor } from '../../../../theme/utils'
 import { getNativeAsset, getNativeCompoundAsset, getToken } from '../../../../util/networks'
 import { formatBigNumber, formatNumber, getBaseTokenForCToken, mulBN } from '../../../../util/tools'
