+++ app/src/components/market/sections/market_bond/market_bond.tsx
@@ -5,7 +5,8 @@ import { RouteComponentProps, withRouter } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { STANDARD_DECIMALS } from '../../../../common/constants'
-import { useConnectedWeb3Context, useContracts, useCpkProxy } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useContracts, useCpkProxy } from '../../../../hooks'
 import { getLogger } from '../../../../util/logger'
 import { getNativeAsset } from '../../../../util/networks'
 import { RemoteData } from '../../../../util/remote_data'
