+++ app/src/components/market/sections/market_pooling/scalar_market_pool_liquidity.tsx
@@ -6,10 +6,10 @@ import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { DOCUMENT_FAQ, STANDARD_DECIMALS } from '../../../../common/constants'
+import { useConnectedWeb3Context } from '../../../../contexts'
 import {
   useCollateralBalance,
   useCompoundService,
-  useConnectedWeb3Context,
   useContracts,
   useCpkAllowance,
   useCpkProxy,
