+++ app/src/components/market/sections/market_pooling/user_pool_data.tsx
@@ -2,7 +2,8 @@ import { BigNumber } from 'ethers/utils'
 import React from 'react'
 import styled from 'styled-components'
 
-import { useCompoundService, useConnectedWeb3Context } from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useCompoundService } from '../../../../hooks'
 import { formatBigNumber, formatNumber, getInitialCollateral } from '../../../../util/tools'
 import { CompoundTokenType, Token } from '../../../../util/types'
 import { TitleValue } from '../../../common'
