+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -6,13 +6,8 @@ import ReactTooltip from 'react-tooltip'
 import styled from 'styled-components'
 
 import { STANDARD_DECIMALS } from '../../../../common/constants'
-import {
-  useAsyncDerivedValue,
-  useCompoundService,
-  useConnectedWeb3Context,
-  useContracts,
-  useSymbol,
-} from '../../../../hooks'
+import { useConnectedWeb3Context } from '../../../../contexts'
+import { useAsyncDerivedValue, useCompoundService, useContracts, useSymbol } from '../../../../hooks'
 import { MarketMakerService } from '../../../../services'
 import { getLogger } from '../../../../util/logger'
 import { getNativeAsset, getWrapToken } from '../../../../util/networks'
