+++ app/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from '../contexts/connectedWeb3'
 export * from './useBalance'
 export { use24hsVolume } from './use24hsVolume'
 export { useAsyncDerivedValue } from './useAsyncDerivedValue'
@@ -14,7 +13,7 @@ export { useDetectAdblocker } from './useDetectAdblocker'
 export { useDocumentDescription } from './useDocumentDescription'
 export { useDocumentTitle } from './useDocumentTitle'
 export { useFundingBalance } from './useFundingBalance'
-export { useGraphMarketMakerData } from './useGraphMarketMakerData'
+export { useGraphMarketMakerData } from './Graph/useGraphMarketMakerData'
 export { useGraphMarketUserTxData } from './useGraphMarketUserTxData'
 export { useGraphMeta } from './useGraphMeta'
 export { useMarketMakerData } from './useMarketMakerData'
